requirements.txt
requests~=2.31.0
flask~=3.0.0
wtforms~=3.1.1
python-dotenv~=1.0.0
flask-wtf~=1.1.1
flask-sqlalchemy~=3.0.3
flask-migrate~=4.0.4
flask-login~=0.6.2
flask-mail~=0.9.1
flask-bcrypt~=1.0.1
psycopg2-binary~=2.9.6
flask-bootstrap~=3.3.7.1
